add local mocha to path
This is a simplistic attempt to find the local path to mocha. This
code makes 2 assumptions:

a) The `files` argument (defaults to `test/`) to `mocha-pipelines` bin
contains a `test/` directory.
b) The project root (where `node_modules/.bin/mocha` lives) can be
located using the parent of the `test/` directory found in a).

If there isn't a locally installed mocha in the project, then
this should fallback to a globally installed mocha (if any).